Add support for vectors on the rhs of linalg.lu_solve
I really don't like this feature, but I guess that's what we are
supporting. I need it to make `linalg.solve` a composition of two
functions (albeit with custom derivative formulas).

Rant: This PR and the next one in this stack show how annoying and
inconvenient is to support NumPy's behaviour
of allowing the rhs to be a vector. We have to hardcode in every formula
(forward AD / backward AD / the function itself / every test) a few lines to handle
that case. This should be done by the user by calling `.unsqueeze(-1)` before calling
this function if they want to operate on column vectors.

ghstack-source-id: 83e023d24035f2cb41120a3c6ca638c11e44e21a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74045